+++ database/migrations/2021_10_13_165947_rename_and_remove_col_seller_wallet.php
@@ -13,10 +13,20 @@ class RenameAndRemoveColSellerWallet extends Migration
      */
     public function up()
     {
-        Schema::table('seller_wallets', function (Blueprint $table) {
+        if (Schema::hasTable('seller_wallets')) {
+
+            if (Schema::hasTable('seller_wallets')) {
+
+
+                Schema::table('seller_wallets', function (Blueprint $table) {
             $table->dropColumn('total_earning');
             $table->renameColumn('balance', 'total_earning');
         });
+
+
+            }
+
+        }
     }
 
     /**
