+++ database/migrations/2021_10_13_170258_rename_and_remove_col_admin_wallet.php
@@ -13,10 +13,20 @@ class RenameAndRemoveColAdminWallet extends Migration
      */
     public function up()
     {
-        Schema::table('admin_wallets', function (Blueprint $table) {
+        if (Schema::hasTable('admin_wallets')) {
+
+            if (Schema::hasTable('admin_wallets')) {
+
+
+                Schema::table('admin_wallets', function (Blueprint $table) {
             $table->dropColumn('inhouse_sell');
             $table->renameColumn('balance', 'inhouse_earning');
         });
+
+
+            }
+
+        }
     }
 
     /**
