+++ database/migrations/2021_10_14_061603_column_update_order_transaction.php
@@ -13,10 +13,20 @@ class ColumnUpdateOrderTransaction extends Migration
      */
     public function up()
     {
-        Schema::table('order_transactions', function (Blueprint $table) {
+        if (Schema::hasTable('order_transactions')) {
+
+            if (Schema::hasTable('order_transactions')) {
+
+
+                Schema::table('order_transactions', function (Blueprint $table) {
             $table->dropColumn('id');
             $table->string('transaction_id')->nullable();
         });
+
+
+            }
+
+        }
     }
 
     /**
