+++ database/migrations/2021_10_16_234037_change_col_type_translation.php
@@ -13,9 +13,19 @@ class ChangeColTypeTranslation extends Migration
      */
     public function up()
     {
-        Schema::table('translations', function (Blueprint $table) {
+        if (Schema::hasTable('translations')) {
+
+            if (Schema::hasTable('translations')) {
+
+
+                Schema::table('translations', function (Blueprint $table) {
             $table->dropColumn('id');
         });
+
+
+            }
+
+        }
     }
 
     /**
